requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pyjwt
passlib
bcrypt
pandas
python-multipart
passlib[bcrypt]